fixed heritage with user

+++ src/main/java/com/example/HealthCare/Repositories/DossierMedicalRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface DossierMedicalRepository extends JpaRepository<DossierMedical,Integer> {
+public interface DossierMedicalRepository extends JpaRepository<DossierMedical,Long> {
     Optional<DossierMedical> findByPatient(Patient patient);
     Page<DossierMedical> findDossierMedicalByDiagnostic(Pageable pageable, String diagnostic);
 }
