+++ src/main/java/com/example/HealthCare/Repositories/MedecinRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MedecinRepository extends JpaRepository<Medecine,Integer> {
+public interface MedecinRepository extends JpaRepository<Medecine,Long> {
     Page<Medecine> findMedecineBySpecialiteContaining(String specialite , Pageable pageable);
 }
