+++ src/main/java/com/example/HealthCare/Repositories/RendezVousRepository.java
@@ -10,9 +10,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface RendezVousRepository extends JpaRepository<RenderVous,Integer> {
-    List<RenderVous> findByPatient_Id(int patientId);
-    List<RenderVous> findByMedecine_Id(int medecineId);
+public interface RendezVousRepository extends JpaRepository<RenderVous,Long> {
+    List<RenderVous> findByPatient_Id(Long patientId);
+    List<RenderVous> findByMedecine_Id(Long medecineId);
 
     Page<RenderVous> findRenderVousByStatut(RendezVousStatutEnum statut, Pageable pageable);
 }
